:star: Hidden 16bit support
Importing 16bit images will currently deprecate the higher resolution
but is supported. For future expanding of functionality to 16bit
it should only require to remove the `dtype=np.uint8` of the
`load_image_to_numpy` function.

Of course it is also needed to give user option to record 16bit
images when starting experiments and extending under the hood
functionality for getting this info to the scanners

+++ scanomatic/image_analysis/analysis_image.py
@@ -178,7 +178,7 @@ class ProjectImage(object):
 
         try:
 
-            self.im = load_image_to_numpy(path, IMAGE_ROTATIONS.Portrait)
+            self.im = load_image_to_numpy(path, IMAGE_ROTATIONS.Portrait, dtype=np.uint8)
             self._im_loaded = True
 
         except (TypeError, IOError):
@@ -189,7 +189,7 @@ class ProjectImage(object):
             self._logger.warning("Failed to load image at '{0}', trying '{1}'.".format(path, alt_path))
             try:
 
-                self.im = load_image_to_numpy(alt_path, IMAGE_ROTATIONS.Portrait)
+                self.im = load_image_to_numpy(alt_path, IMAGE_ROTATIONS.Portrait, dtype=np.uint8)
                 self._im_loaded = True
 
             except (TypeError, IOError):
